+++ Tests/New-PASRequest.Tests.ps1
@@ -53,6 +53,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 			$response = $InputObj | New-PASRequest
 
 		}
+
 		Context 'Mandatory Parameters' {
 
 			$Parameters = @{Parameter = 'AccountID' }
@@ -61,13 +62,11 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 				param($Parameter)
 
-				(Get-Command New-PASRequest).Parameters["$Parameter"].Attributes.Mandatory | Should -Be $true
+				(Get-Command New-PASRequest).Parameters["$Parameter"].Attributes.Mandatory | Select-Object -Unique | Should -Be $true
 
 			}
 
 		}
-
-
 
 		Context 'Input' {
 
@@ -133,6 +132,61 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 				$Script:ExternalVersion = '1.0'
 				{ $InputObj | New-PASRequest } | Should -Throw
 				$Script:ExternalVersion = '0.0'
+			}
+
+			It 'has request body with expected BulkItems property' {
+				$Requests = $InputObj | New-PASRequestObject
+				New-PASRequest -BulkItems $Requests
+
+				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter {
+
+					$Script:RequestBody = $Body | ConvertFrom-Json
+
+					($Script:RequestBody).BulkItems -ne $null
+
+				} -Times 1 -Exactly -Scope It
+
+
+			}
+
+			It 'has BulkItems property with expected format' {
+
+				$Script:RequestBody.BulkItems.Operation | Should -Be 'Add'
+
+			}
+
+			It 'has request body with expected SearchParam' {
+
+				New-PASRequest -Search SomeSearchTerm
+
+				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter {
+
+					$Script:RequestBody = $Body | ConvertFrom-Json
+
+					($Script:RequestBody).BulkFilter -ne $null
+
+				} -Times 1 -Exactly -Scope It
+			}
+
+			It 'has FilterParams property with expected format' {
+				$Script:RequestBody.BulkFilter.FilterParams.SearchParam.Search | Should -Be 'SomeSearchTerm'
+			}
+
+			It 'has request body with expected AccountsFilters' {
+
+				New-PASRequest -SavedFilter Favorites
+
+				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter {
+
+					$Script:RequestBody = $Body | ConvertFrom-Json
+
+					($Script:RequestBody).BulkFilter -ne $null
+
+				} -Times 1 -Exactly -Scope It
+			}
+
+			It 'has FilterParams property with expected format' {
+				$Script:RequestBody.BulkFilter.FilterParams.AccountsFilters.SavedFilter | Should -Be 'Favorites'
 			}
 
 		}